restructure go application  to modular approach
