+++ app/Dockerfile
@@ -1,26 +1,21 @@
-# Start from the official Go image for building
 FROM golang:1.24 AS builder
 
-# Set the working directory inside the container
 WORKDIR /app
 
-# Copy go source files
+COPY go.mod go.sum ./
+RUN go mod download
+
 COPY . .
 
-# Build the Go application
 RUN go build -o main .
 
-# Final slim image to run the app
+# stage 2 build: put only required stuff
 FROM debian:bookworm-slim
 
-# Set working directory in final image
 WORKDIR /app
 
-# Copy the compiled binary from builder stage
 COPY --from=builder /app/main .
 
-# Expose port 8080 to match containerPort in deployment
 EXPOSE 8080
 
-# Run the binary
 CMD ["./main"]
