+++ app/internal/handlers/handlers.go
@@ -1,0 +1,16 @@
+package handlers
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func Home(w http.ResponseWriter, r *http.Request) {
+	fmt.Printf("Request received from IP: %s\n", r.RemoteAddr)
+	fmt.Fprintln(w, "Hello from Go!")
+}
+
+func Health(w http.ResponseWriter, r *http.Request) {
+	w.WriteHeader(http.StatusOK)
+	fmt.Fprintln(w, "OK")
+}
