+++ app/internal/metrics/metrics.go
@@ -1,0 +1,25 @@
+package metrics
+
+import (
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+)
+
+var (
+	RequestCounter = promauto.NewCounterVec(
+		prometheus.CounterOpts{
+			Name: "myapp_http_requests_total",
+			Help: "Total number of HTTP requests",
+		},
+		[]string{"path", "status"},
+	)
+	
+	RequestDuration = promauto.NewHistogramVec(
+		prometheus.HistogramOpts{
+			Name:    "myapp_http_request_duration_seconds",
+			Help:    "Duration of HTTP requests in seconds",
+			Buckets: prometheus.DefBuckets,
+		},
+		[]string{"path"},
+	)
+)
