+++ app/internal/middleware/response_writer.go
@@ -1,0 +1,21 @@
+package middleware
+
+import "net/http"
+
+type responseWriter struct {
+	http.ResponseWriter
+	statusCode int
+}
+
+func NewResponseWriter(w http.ResponseWriter) *responseWriter {
+	return &responseWriter{w, http.StatusOK} 
+}
+
+func (rw *responseWriter) WriteHeader(code int) {
+	rw.statusCode = code
+	rw.ResponseWriter.WriteHeader(code)
+}
+
+func (rw *responseWriter) StatusCode() int {
+	return rw.statusCode
+}
